requirements.txt
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
psycopg==3.2.9
dj-database-url==3.0.1
whitenoise==6.7.0
gunicorn==23.0.0
httpx==0.27.2
openpyxl==3.1.5
reportlab==4.2.5
